Split remaining class and enum in Utils.cs into Category.cs and MachineState.cs (khanshoaib3/stardew-access#112)
