+++ stardew-access/Features/StaticTiles.cs
@@ -1,5 +1,6 @@
 using StardewValley;
 using System.Text.Json;
+using stardew_access.Utils;
 using static stardew_access.Utils.JsonLoader;
 
 namespace stardew_access.Features
