+++ stardew-access/Utils/MachineState.cs
@@ -1,0 +1,11 @@
+//using System;
+//using System.Collections.Generic;
+//using System.Drawing;
+
+namespace stardew_access.Utils
+{
+    public enum MachineState
+    {
+        Ready, Busy, Waiting
+    }
+}
